- Changed BootstrapLayoutGenerator CDI scope - Refactor to help creating different implementations
